Created "old_release_TPA_v3.js"
The script is named old because a new one is coming soon! Stay tuned.

You can use this version for now, it is fully working.

Note: It may contain a couple bugs as it's not perfect, but it works as intended, so don't be afraid and give it a try!