What the... "x"?
Wanted a script that will contain every update in it? Here it is, take it or leave it.
The most relevant version (apparently 3.2 which is not out yet) will be straight in this file, and will get updated right here, without requiring you to search for a new file in the repo.
This does not mean that every new version will erase the previous one! When an update comes out, it is going to be pushed into TWO files: this one, and a new one, related to the update itself. So... guess, that's a win-win for everyone, right?